refactor(models): 클래스를 models로 이동
- 파일과 디렉토리 구조를 명확히 하고 가독성 향상을 위해 Car과 Race
  클래스를 models 디렉토리 하위로 이동
